+++ runtimes/peregrine/src/runtime_apis.rs
@@ -17,11 +17,10 @@ use sp_api::impl_runtime_apis;
 use sp_core::OpaqueMetadata;
 use sp_inherents::{CheckInherentsResult, InherentData};
 use sp_runtime::{
-	create_runtime_str,
 	traits::{Block as BlockT, TryConvert},
 	ApplyExtrinsicResult, KeyTypeId,
 };
-use sp_version::RuntimeVersion;
+use sp_version::{ApisVec, RuntimeVersion};
 
 use kilt_runtime_api_did::RawDidLinkedInfo;
 use kilt_support::traits::ItemFilter;
@@ -47,8 +46,12 @@ use crate::{
 	parachain::ConsensusHook,
 	xcm::UniversalLocation,
 	AssetSwitchPool1, Aura, Block, DotNames, Executive, InherentDataExt, ParachainStaking, ParachainSystem, Runtime,
-	RuntimeCall, RuntimeGenesisConfig, SessionKeys, TransactionPayment, UniqueLinking,
+	RuntimeCall, RuntimeGenesisConfig, SessionKeys, TransactionPayment, UniqueLinking, VERSION,
 };
+
+// This is necessary since by default `RUNTIME_API_VERSIONS` generated by
+// `impl_runtime_apis` is private.
+pub(crate) const RUNTIME_API_VERSION: ApisVec = RUNTIME_API_VERSIONS;
 
 impl_runtime_apis! {
 	impl sp_api::Core<Block> for Runtime {
@@ -474,28 +477,5 @@ impl_runtime_apis! {
 			build_config::<RuntimeGenesisConfig>(config)
 		}
 
-	}
-}
-
-/// This runtime version.
-#[sp_version::runtime_version]
-pub const VERSION: RuntimeVersion = RuntimeVersion {
-	spec_name: create_runtime_str!("mashnet-node"),
-	impl_name: create_runtime_str!("mashnet-node"),
-	authoring_version: 4,
-	spec_version: 11500,
-	impl_version: 0,
-	apis: RUNTIME_API_VERSIONS,
-	transaction_version: 8,
-	state_version: 0,
-};
-
-/// The version information used to identify this runtime when compiled
-/// natively.
-#[cfg(feature = "std")]
-pub fn native_version() -> sp_version::NativeVersion {
-	sp_version::NativeVersion {
-		runtime_version: VERSION,
-		can_author_with: Default::default(),
 	}
 }
